adding tokens on apps creation
